Fix frame skipping when flies extend their legs to far
To minimize the interference of leg extensions on the wingbeat tracking
the dark legs are thresholded away and replace with nans. If a leg
covers the entire radial distance of the wingbeat mask, the following
code raises an Exception which is apparently not fast enough for keeping
up with the incoming frames.

This patch replaces the thresholded leg areas with the average
background intensity and fixes the frame skipping issue.